requirements.txt
streamlit==1.21.0
google-generativeai==0.1.0 
